+++ src/PSOTLP/Exporters/OTLPTraceExporter.cs
@@ -42,7 +42,7 @@ namespace PSOTLP.Exporters
             }
 
             var definition = OTLPEndpointRegistry.Get(OTLPEndpointRegistry.ExportTracesName);
-            var uri = OTLPUriBuilder.Build(connection.EndpointUri, definition, connection.TracesEndpointUri);
+            var uri = OTLPUriBuilder.Build(connection.EndpointUri, definition, connection.TracesEndpointUri, connection.Encoding);
             var payload = BuildPayload(connection, spans);
 
             var body = _serializer.SerializeTraces(payload);